server/src/requirements.txt
aiohttp==3.7.4.post0
motor==2.3.1
pymongo==3.11.3
requests==2.25.1
python-engineio==2.3.2
python-socketio==2.0.0
aioredis==1.3.1
pytz==2021.1